// src/commons/interfaces/GeDocumental.ts
export interface IAdmisiones {
  Apellido1: string;
  Nombre2: string;
  Apellido2: string;
  Nombre1: string;
  Consecutivo: string;
  SubCentro: number;
  TipoID: string;
  IdPaciente: string;
  Ordina: number;
  CodigoEntidad: string;
  PolizaNo: string;
  Confirmado: string;
  CedulaResponsable: string;
  NombreResponsable: string;
  DireccionResponsable: string;
  TelefonoResponsable: string;
  AutorizacionNo: string;
  FechaCreado: string;
  FechaModificado: string;
  CreadoPor: number;
  ModificadoPo: number;
  Estad: number;
  FechaSalida: string;
  TipoAdmisio: number;
  SalidaPo: number;
  Habitacion: string;
  FechaHospital: string;
  TipoCirugi: number;
  MotivoConsulta: string;
  Facturad: number;
  FacturadoPo: number;
  FacturaNo: string;
  MDTratante: string;
  Especialidad: string;
  Anulad: number;
  Procedimiento1: string;
  Procedimiento2: string;
  Procedimiento3: string;
  Procedimiento4: string;
  Procedimiento5: string;
  Procedimiento6: string;
  TipoEvent: number;
  EmpresaPaciente: string;
  MunicipioEmpresa: string;
  NMDTratante: string;
  VrRecibido: string;
  Mpag: number;
  SemanasCot: string;
  Nive: number;
  Capitad: number;
  NumeroPacient: number;
  MotivoAnulacion: string;
  AnuladoPo: number;
  FechaAnulado: string;
  Ruta: string;
  VrInicia: number;
  VrConsumid: number;
  Radicado: string;
  Diag1: string;
  Diag2: string;
  Diag3: string;
  FechaAutorizada: string;
  FCreada: string;
  ConceptoRC: string;
  NumeroRC: string;
  FechaImpreso: string;
  ListoFarmaci: number;
  ListoLaboratorio: number;
  EvitarDelet: number;
  Examene: number;
  CCos: number;
  EDelete: number;
  SeccionFactura: string;
  ObservacionesEnf: string;
  IdEven: number;
  Impresione: number;
  ImpresionFormato: string;
  Diario: string;
  Remitid: number;
  RemitidoPor: string;
  CodigoRemite: string;
  MpioRemite: string;
  Triag: number;
  Cobertur: number;
  OrigenAtt: number;
  Desti: number;
  TipoSolicitud: string;
  Priorida: string;
  PteAtendid: number;
  HoraAtencion: string;
  IdCi: number;
  Murgencias: string;
  HoraLlegada: string;
  SegundosAtencio: number;
  IDAnarlab: string;
  TratarEnfBas: number;
  tRegim: number;
  TipoAfiliado: string;
  MedicoOrdena: string;
  IPSOrdena: string;
  EspMDOrde: number;
  Prefijo: string;
  selectedHallazgo: string;
  TiposDocumentos: string[];
  tipoDocumento: string;
  // Campos ZeusSalud
  NombreCompleto?: string;
  NombreEntidad?: string;
  NumeroPaciente?: number;
  TipoAfiliacion?: number;
  FechaNacimiento?: string;
  Telefono?: string;
  CorreoE?: string;
  SexoPaciente?: string;
  tRegimen?: string;
  ContratoAlias?: string;
  ContratoRegimen?: number;
}